+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -273,6 +273,7 @@
 /// KernelSpace quirks.
 ///
 #define OC_KERNEL_QUIRKS_FIELDS(_, __) \
+  _(INT64                       , SetApfsTrimTimeout          ,     , FALSE  , ()) \
   _(BOOLEAN                     , AppleCpuPmCfgLock           ,     , FALSE  , ()) \
   _(BOOLEAN                     , AppleXcpmCfgLock            ,     , FALSE  , ()) \
   _(BOOLEAN                     , AppleXcpmExtraMsrs          ,     , FALSE  , ()) \