+++ Changelog.md
@@ -12,6 +12,7 @@ OpenCore Changelog
 - Fixed compatibility with page protection for all binaries
 - Fixed crashes in OpenUsbKbDxe when handling unsupported devices
 - Removed `HdaCodecDump` application in favor of `SysReport`
+- Added `SetApfsTrimTimeout` to tune APFS trim command
 
 #### v0.6.5
 - Fixed installing OpenDuet on protected volumes